Adapting to interpretation of (x:T) now activating scope for T if any.
These were situations where the cast was needed to hint type class
inference but the body was expected to be interpreted in the
type-class (current) scope rather than in the more specialized scope
bound to the cast type.

We add a "%mc" to force using the type-class based generic
interpretation.